// src/app/models/codeMethodeStatisticDataTreeDTO.ts
export class CodeMethodeStatisticDataTreeDTO{  

    methodeName: string;
    mcCabeComplexity: string;
    linesOfCode: string;
    returnQty: string;
    loopQty: string;
    comparisonsQty: string;
    tryCatchQty: string;
    stringLiteralsQty: string;
    numbersQty: string;
    variablesQty: string;
    maxNestedBlocks: string;
    logStatementsQty: string;
}